Adding alias support
This adds support for a single alias and is designed for situations where a project which is run as 'root' (i.e. /) in development is deployed as one of many services (i.e. /service).

This support is therefore limited to single alias (this isn't indended to meet the needs of deployment).