Adding SSL chain file support
Apache doesn't send the entire certificate chain even if other certificates are included in the certificate file (i.e. concatenated). Instead Apache requires a certificate chain file to be used (which can be the same as the certificate file).

This change adds support for this directive and by default will just use the same certificate file.