Switching to include all config files in config-enabled
This removes the need for using lineinfile (which was very temperamental) and simplifies things significantly. To include a configuration file just ensure it is in 'conf-enabled' (sym-linked from 'conf-available').

This approach is scalable and leverages Apache's built in logic for loading a directory of config files, rather than just essentially doing the same thing but worse.

This is a breaking change and module roles will need to be updated to work with this new version.