Adding priminary testing support
This is clearly a proof of concept. I want to significantly improve this by creating, for example, 5 VMs each configured slightly differently (one using SSL, one using no HTTP, etc.).

This will take some time to figure out of course and may end-up being impractical but should at least be tried.

Currently support is added for creating an SSL configured Apache instance both in a Vagrant VM and Terraform DO VM.

The provisioning of these VMs is purposely designed to replicate how this role would be included/used in other projects. This also reduces the need to re-invent the wheel.