Adding support for uploading SSL certificates and associated private key
Private keys were previously uploaded by the core role and certificates were left to the user to provide (i.e. upload somewhere in /app).

To follow best practice and remove the need for the core role to care about SSL this has been refactored into this role. To allow for compatibility with other web-servers variables in this role first check for a set of generic variables. This means if you set the generic variables they will be picked up by any compatible web-server role.

The testing playbooks and group-vars have been updated to reflect these changes, and are therefore simplier.

These new variables need to be documented prior to release. The core role and other web-server roles still need to be updated to follow this new convention.