Separating virtualhost configuration and SSL configuration
This is part of a wider process to split up things related to configuring Apache generally and things for configuring the default virtual-hosts (http/https) specifically.

This is to allow for supporting disabling one or both of these defaults in cases where only http or https connections should be allowed, or where custom virtualhosts are desired, in which case this role should only focus on configuring Apache itself.